src/requirements.txt
astroid==2.11.2
autopep8==1.5.7
beautifulsoup4==4.9.3
boto3==1.17.95
botocore==1.20.112
bs4==0.0.1
certifi==2024.7.4
chardet==4.0.0
dill==0.3.4
idna==3.7
influxdb==5.3.1
influxdb-client==1.29.1
isort==5.8.0
jmespath==0.10.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
msgpack==1.0.2
platformdirs==2.5.1
paho-mqtt==2.1.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.32.4
Rx==3.2.0
s3transfer==0.4.2
six==1.16.0
soupsieve==2.2.1
splunk-hec-handler==1.1.0
toml==0.10.2
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.19
wrapt==1.12.1
